api: move plotting to its own module
