+++ qcodes/plotting/__init__.py
@@ -1,0 +1,15 @@
+from .matplotlib_helpers import (
+    apply_auto_color_scale,
+    apply_color_scale_limits,
+    auto_color_scale_from_config,
+    find_scale_and_prefix,
+)
+from .plotting import auto_range_iqr
+
+__all__ = [
+    "apply_auto_color_scale",
+    "apply_color_scale_limits",
+    "auto_color_scale_from_config",
+    "auto_range_iqr",
+    "find_scale_and_prefix",
+]
